Put all commands together
