+++ cmd/skuba/node/remove.go
